requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.4
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
schedule==1.2.1
